requirements.txt
assemblyai==0.18.0
moviepy==1.0.3
Pillow==10.0.1
playsound==1.3.0
praw==7.7.1
proglog==0.1.10
pydub==0.25.1
Requests==2.31.0
tqdm==4.66.1
